requirements.txt
fastapi
uvicorn[standard]
yt-dlp
boto3
python-dotenv
librosa
noisereduce
scipy
